prevent name clashes in commands, close #3
To merge TypeScript interface, the source code must be
treated as a script. But this means all local variables inside
the commands.ts are going to clash with any other variables
with same names in other TS files.

Here is an ugly workaround. Put your variables in commands.ts
in a closure, but leave the `declare interface ...` block outside.

    ;(() => {
      const { foo, foo2 } = require('./actions')
      const _ = Cypress._
      Cypress.Commands.add('foo', foo)
      Cypress.Commands.add('foo2', foo2)
    })()
    declare namespace Cypress {
      interface Chainable {
        // note we don't have access to actual foo and foo2
        // so must duplicate the type signatures
        foo: () => string
        foo2: () => string
      }
    }

You have to do the same if you want to use `_` in the spec files

    ;(() => {
      // try to have name constant in test as in commands.ts
      const _ = Cypress._
      it('works', () => {
        expect(Cypress.version).to.be.a('string')
      })
    })()

If there is a better solutio, I would like to know it